Extract utility methods.
